% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_glm.R
\name{get_predicted}
\alias{get_predicted}
\title{Get Predicted Values from a Model}
\usage{
get_predicted(mod, exclude_cols = NULL, type = "link")
}
\arguments{
\item{mod}{A fitted model object, such as from \code{glm}, \code{lm}, or \code{robust_glm}.}

\item{exclude_cols}{Columns in \code{mod$data} to exclude from transformation. A
tidyselect specification.}
}
\value{
A tibble with linear predictors, predicted outcome, and standard
errors of the predicted outcome.
}
\description{
This function calculates predicted values from a fitted model, setting covariates
to their median (for numeric variables) or mode (for categorical variables).
}
