% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_props.R
\name{get_props}
\alias{get_props}
\title{Compute Proportions}
\usage{
get_props(
  data,
  id,
  cluster,
  supercluster = NULL,
  by = NULL,
  add_supercluster_prop = F,
  add_other_cols = T
)
}
\arguments{
\item{data}{Data frame containing \code{id} and \code{cluster}.}

\item{id}{Character. Column name representing the observations identifier.}

\item{cluster}{Character. Column name representing the cluster identifier.}

\item{supercluster}{Character. Column name representing the supercluster
identifier. Optional.}

\item{by}{Character. Column name(s) of grouping variables, a.k.a.
denominators for proportions. Optional.}

\item{add_supercluster_prop}{Logical indicating whether to include
proportions per id across superclusters.}

\item{add_other_cols}{Logical indicating whether to summarize other columns of \code{data} by \code{c(id, supercluster, by)}. Numeric columns are summarized
to their mean, factor to the reference level and character to the mode.}
}
\value{
A long data frame with computed proportions per group and summary
statistics.
\itemize{
\item \code{"num_cluster"}: The number of occurrences of each group combination.
\item \code{"props"}: The proportions of each group combination within its
supercluster.
\item \code{"supercluster"}: The supercluster associated with each group.
}
}
\description{
This function computes proportions per id across cluster. It can group by
supercluster and compute proportions per group(s). It can also add
supercluster proportions across all its levels. Finally, it summarizes
other columns of data per grouping columns.
}
\details{
This function calculates proportions per group based on the counts of another
column, and summarizes other columns by group, calculating means for numeric columns,
reference levels for factor columns, and the mode for character columns.
}
\examples{
df <- data.frame(id = rep(1:3, each = 8),
                 cluster = rep(c("C1", "C2"), each = 4, times = 3),
                 supercluster = rep(c("S1", "S2"), each = 4, times = 3),
                 value1 = runif(24),
                 value2 = as.factor(LETTERS[1:12]),
                 value3 = LETTERS[1:12],
                 value4 = NA)
df
get_props(df, "id", "cluster", "supercluster", add_supercluster_prop = T)
get_props(df, "id", "cluster", "supercluster")
get_props(df, "id", "cluster")

}
