% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy_model}
\alias{tidy_model}
\title{Tidy Model Summary Statistics}
\usage{
tidy_model(model, id = NULL, ...)
}
\arguments{
\item{model}{A list of or a single robust_glm object.}

\item{id}{A character string specifying the identifier column name
(default is "id").}

\item{...}{Arguments passed to \code{broom::tidy()}.}
}
\value{
A data frame containing tidied model summary statistics with
columns:
\describe{
\item{\code{id}}{Identifier column name.}
\item{\code{converged}}{Convergence status of the model.}
\item{\code{dispersion}}{Dispersion value of the model.}
\item{\code{glance}}{Summary statistics of the model.}
\item{\code{rmse}}{Root mean square error.}
\item{\code{mse}}{Mean square error.}
\item{\code{hosmer_chisq}}{Hosmer goodness-of-fit test statistic.}
\item{\code{hosmer_df}}{Degrees of freedom for Hosmer goodness-of-fit
test.}
\item{\code{hosmer_p.value}}{p-value of Hosmer goodness-of-fit test.}
\item{\code{R2_Tjur}}{Tjur's R-squared.}
\item{\code{Log_loss}}{Logarithmic loss.}
}
}
\description{
Tidies summary statistics of model objects, such as convergence status,
dispersion, and performance metrics.
}
