% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy_terms}
\alias{tidy_terms}
\title{Tidy Model Terms}
\usage{
tidy_terms(model, id = NULL, exponentiate = F)
}
\arguments{
\item{model}{A list of or a single robust_glm object.}

\item{id}{A character string specifying the identifier column name
(default is "id").}

\item{exponentiate}{Logical. Exponentiate estimate and confidence
intervals.}
}
\value{
A data frame containing tidied model terms with columns:
\describe{
\item{\code{id}}{Identifier column name.}
\item{\code{term}}{Model term names.}
\item{\code{estimate}}{Estimate of coefficients.}
\item{\code{std.error}}{Standard error of coefficients.}
\item{\code{statistic}}{Value of test statistics.}
\item{\code{p.value}}{p-value of test statistics.}
\item{\code{conf.low}}{Lower bound of confidence interval.}
\item{\code{conf.high}}{Upper bound of confidence interval.}
\item{\code{std.error_hc}}{Sandwich standard error of coefficients.}
\item{\code{statistic_hc}}{Value of test statistics with sandwich
standard errors.}
\item{\code{p.value_hc}}{p-value of test statistics with sandwich
standard errors.}
\item{\code{conf.low_hc}}{Lower bound of confidence interval with
sandwich standard errors.}
\item{\code{conf.high_hc}}{Upper bound of confidence interval with
sandwich standard errors.}
\item{\code{estimate_exp}}{Exponentiated estimate.}
\item{\code{conf.low_exp}}{Lower bound of exponentiated confidence
interval.}
\item{\code{conf.high_exp}}{Upper bound of exponentiated confidence
interval.}
\item{\code{std.error_exp}}{Exponentiated standard error.}
\item{\code{std.error_hc_exp}}{Exponentiated sandwich standard error.}
}
}
\description{
This function tidies model terms, including coefficients, confidence intervals,
and sandwich standard errors, from a list of model objects or a single model object.
}
