% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_props.R
\name{get_props}
\alias{get_props}
\title{Compute Proportions of Clustered Observations}
\usage{
get_props(data, id, cluster, prop_by = NULL, split_by = NULL)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{id}{A string specifying the column name of the unique sample identifier.}

\item{cluster}{A string specifying the column name of the cluster identifier.}

\item{prop_by}{(Optional) A string specifying columns to group by when computing proportions. Columns to fill with explicit zeroes.}

\item{split_by}{(Optional) A vector of strings specifying grouping columns
for which separate proportion calculations should be performed. Columns
not to fill with explicit zeroes.}
}
\value{
A data frame with cluster proportions and counts per unique combination
of the grouping variables.
}
\description{
This function computes the proportion of observations within each cluster,
optionally stratified by additional grouping variables.
}
\examples{
library(dplyr)

# Example dataset
set.seed(42)
data <- data.frame(
  id = rep(1:5, each = 4),
  cluster = sample(letters[1:3], 20, replace = TRUE),
  prop_by = rep(c("X", "Y"), each = 10),
  split_by = rep(c("A", "B"), times = 10)
)

# Basic usage
get_props(data, id = "id", cluster = "cluster")

# Add explicit zeroes to all levels of "prop_by"
get_props(data, id = "id", cluster = "cluster", prop_by = "prop_by")

# Ensure no new combinations outside "split_by" are added
get_props(data, id = "id", cluster = "cluster", split_by = "split_by")

# It can also be integrated into piped commands. Grouped dataframes are
treated as split_by.
data \%>\%
  group_by(split_by) \%>\%
  get_props(id = "id", cluster = "cluster", prop_by = "prop_by")

}
